TST: move server test fixtures to conftest
